// src/lib/enums/filters.ts
export enum SORT_DIRECTION {
  ASC = 'asc',
  DESC = 'desc',
}

export enum SORT_BY {
  DUE_DATE = "dueDate",
  PRIORITY = "priority",
}

export enum SORT_OPTION {
  NONE = "none",
  DUE_DATE_ASC = "dueDateAsc",
  DUE_DATE_DESC = "dueDateDesc",
  PRIORITY_ASC = "priorityAsc",
  PRIORITY_DESC = "priorityDesc",
}